cgame: moved the hud load & save methods to their own file
